requirements.txt
discord.py
winrt
requests